updates after lance feedback
