+++ Makefile
@@ -17,14 +17,25 @@ export PRINT_HELP_PYSCRIPT
 help: 
 	@python -c "$$PRINT_HELP_PYSCRIPT" < $(MAKEFILE_LIST)
 
+clean: ## Cleanup all the things
+	rm -rf .tox
+	rm -rf .pytest_cache
+	rm -rf .coverage
+	rm -rf *.egg-info
+	rm -rf build
+	rm -rf dist
+	rm -rf htmlcov
+	find . -name '*.pyc' | xargs rm -rf
+	find . -name '__pycache__' | xargs rm -rf
+
 docker: ## build docker container for testing
 	@echo "Building test env with docker-compose"
 	docker-compose -f docker/docker-compose.yml build assassin
 	@docker-compose -f docker/docker-compose.yml run assassin /bin/bash
 
 python: ## setup python3
-	if [ -f 'python/requirements.txt' ]; then pip3 install -rpython/requirements.txt; fi
+	if [ -f 'requirements.txt' ]; then pip3 install -rrequirements.txt; fi
 
 test: python ## run tests in container
-	if [ -f 'python/requirements-test.txt' ]; then pip3 install -rpython/requirements-test.txt; fi
-	cd python && tox
+	if [ -f 'requirements-test.txt' ]; then pip3 install -rrequirements-test.txt; fi
+	tox
