+++ README.md
@@ -1,2 +1,5 @@
 # Assassin
+
 What can a potential attacker know about you without being alerted to their recon?  Assassin can help you find out.
+
+![Assassin](https://github.com/hotpeppersec/Assassin/blob/master/Assassin.png)
