+++ SECURITY.md
@@ -1,0 +1,16 @@
+# Security Policy
+
+- This project uses lgtm.com to scan for security issues.
+- You can [click here to follow a link to the most recent security scan report](https://lgtm.com/projects/g/hotpeppersec/Assassin/alerts/?mode=list).
+
+## Supported Versions
+
+All versions of this project are currently being supported with security updates.
+
+| Version | Supported          |
+| ------- | ------------------ |
+| 0.1     | :white_check_mark: |              |
+
+## Reporting a Vulnerability
+
+Use [the "issues" tab in this repo](https://github.com/hotpeppersec/Assassin/issues) to report any possible security issues to the authors.
