+++ assassin/validator.py
